Meta Desk Configuration of languages
Allows "slightly" easier addition of languages. 
The parsers still need to be added both to the source structure and buildsuper scripts. 
You also need to rename the capture's to the theme colours (i want to add some kind of mapping to the mdesk file for this).
The metadesk build step, generates the jpts_register_languages function which initialises the languages, as well as concatenating any queries for extended languages (i.e. typescript extends js and cpp extends c)